+++ doc/changelog.rst
@@ -9,6 +9,7 @@ Features
 
 - Added compute node monitoring and graphs to GUI and API - `#13 <https://github.com/erigones/esdc-ce/issues/13>`__
 - Added ``cpu_type`` parameter into vm_define API call - `#76 <https://github.com/erigones/esdc-ce/issues/76>`__
+- Updated metadata input fields to accept raw JSON input `#79 <https://github.com/erigones/esdc-ce/issues/79>`__
 
 Bugs
 ----